Wake-up the application after successfull connect
Note: this commit is based on the modifications done in mptcp_set_state from
4504ac7498ec0be6f47e6e0abf98d80d7fe8ea81.

The stack call for the connect is:
connect(meta-socket) -> inet_stream_connect(meta) -> __inet_stream_connect(meta) ->
-> sk->sk_prot->connect(meta) -> mptcp_v4_connect -> inet_stream_connect(master)->
__inet_stream_connect(master)-> tcp_v4_connect(master).
The problem is that the application blocks in  __inet_stream_connect on the master
socket waiting for a signal in inet_wait_for_connect. This signal should come from
the callback sk_state_change. In order to solve this, we wake-up first the master-socket
thread then we avoid to block for the meta-socket.

In mptcp_add_sock we modify the sk_state_change callback of the master socket to
mptcp_set_state. In regular TCP this callback is used to wake-up the application,
for example after the 3-WAY handshake is finished we should wake up the application.
For mptcp, we don't want to wake the application every time a new handshake for a
subflow is finished but still want to wake it up when the handshake for the master
socket is done. In order to do that in mptcp_set_state, the callback of the
meta-socket is used on the master_socket. This callback is still the unmodified
callback(sock_def_wakeup) from sock_init_data. In this way, the thread waiting in
inet_wait_for_connect on the master-socket is woken-up.

In mptcp_set_state we also set the state of the meta-socket to TCP_ESTABLISHED
so we won't block in __inet_stream_connect on the __meta__ socket.

In the old mptcp_arhitecture the master-socket was created as a clone of the meta-socket
so the sk_wq queue of the meta-socket (queue of threads which wait for data on a specific
socket) was copied to the sk_wq of the master socket.  The new arhitecture removes the
clone part so we don't need this copy of the sk_wq from the meta-socket.

This modification works when the connection finishes successfully. We should also
consider the case when a RST is received from the server.