Propagate the reset upwards to the application
If the connect is unssuccessfull (for example, when we receive RST+ACK in
tcp_rcv_synsent_state_process) the "sk_error_report" will be called on the master-socket.
For the mptcp case, this callback is set to mptcp_sock_def_error_report so we simply check
there if we are still in SYN_SENT state from the tcp state machine(this means that we
have not received a SYN+ACK which could move us to the TCP_ESTABLISHED state).

The resources allocated for mptcp will be freed in mptcp_close.
(__inet_stream_connect->sk_prot->disconnect....)