Partial redesign of the receive path using tcp_read_sock
[Work in progress: This patch still need some work to solve the locking
problems, direct copy and the handling of the fin flag, ...]

The original implementation for the receive path is pretty complex. Packets
received on the sub-flows are enqueued to the meta backlog or prequeue
queues. After processing the meta queues, packets are put in the sub-flow
receive queue. Packets can be copied to user-space only after they have been
put in the meta-receive queue so they are transferred from the sub-flow to
meta receive queue.

In order to simplify the flow we propose to make use of the sub-flow queues.
This implies that the backlog and the prequeue of the meta socket is no longer
used but the backlog of the sub-flow (as future work we need to find a solution
for using the prequeue of the sub-flow). At processing time, packets are
transferred from sub-flow backlog to sub-flow receive queue without using
the meta queues. Another implication is that, in tcp_v4_rcv(), the lock is not
taken on the meta socket but on the sub-flow socket. This should allow to
remove the MPTCP code related to locking from tcp_v4_rcv().

In order to pass the data from the sub-flow level to meta level we defined a
new MPTCP method for the meta-socket based on the tcp_read_sock function - a
standard Kernel API (or at least used by other important users such as NFS and
CIFS). The role of tcp_read_sock function is to extract the skbs from the
receive queue of the subflow sockets and push them to the receive-queue/out of
order queue of the meta-socket.

The DSS mapping is build in the sk_data_ready callback which wakes up the
application only when we have a mapping for skbs from the receive queue of
the subflow.

Signed-off-by: Doru Gucea <doru.gucea@intel.com>