mptcp: Add ADD_ADDR2 option
ADD_ADDR2 implementation adds the truncated HMAC value of the
advertised information as a field in the ADD_ADDR option. The
rationale for this is to prevent injection of unauthorized
ADD_ADDR2 packets to perform a connection hijacking.

ADD_ADDR2 is used only if MPTCP version 1 is agreed by both
participants in the initial handshake. If version 0 is agreed,
ADD_ADDR is used instead guaranteeing retro-compatibility.

This commit also includes the added support for port advertisement.

This implementation complies with RFC6824bis-04.

Signed-off-by: Fabrizio Demaria <fabrizio.demaria@intel.com>