mptcp: Make 'mptcp_hmac_sha1' more flexible
This implementation of the hashing function permits to pass an
arbitrary number of pointers pointing to all the byte strings
composing the complete HMAC message.
The HMAC keys are still fixed and they must be exactly two, of
8 bytes each.

The new function allows to better handle new cases in which HMAC
computation is needed in MPTCP, for example ADD_ADDR2.

IPv6 HMAC computation is also updated.

Signed-off-by: Fabrizio Demaria <fabrizio.demaria@intel.com>