mptcp: Refactor input path to use tcp_read_sock()
The current way of moving skbs from subflow to meta socket involves
manually iterating through the socket receive queue.  The preferred way
to do this upstream is to use the existing tcp_read_sock() API.

Using tcp_read_sock() attracts a few corner cases, however.  First off,
DSS mapping detection has to be done as soon as an skb gets added to the
receive queue.  To do this, we use the previously introduced callback.
We need to clone skbs instead of passing them directly to the meta sock,
because tcp_read_sock() handles the freeing of skbs from the subflow.

In the case of a DSS-carying skb filling in a hole in the ofo queue at
the subflow level, additional skbs may be added before sk_data_ready()
is called on the subflow.  One or more of these skbs can contain another
DSS mapping, thus in this case a subflow may have, for a brief moment,
more than one mapping pending.  To alleviate this, we introduce a
mapping structure and instead hold a list of such mappings in the
socket.  To minimize the overhead, a separate slab is used for mappings.

Another aspect that further complicates the issue is that, using
tcp_read_sock(), the actor must take the decision whether to consume the
skb before moving on to the next one.  However, in mptcp, we need to
checksum a whole DSS mapping and throw away its contents if the checksum
is bad.  We achieve this by moving skbs to a separate queue and only
checksumming them when the mapping is complete.  It is important to do
this before tcp_read_sock() returns, otherwise we would send an ACK with
the new subflow sequence number and the old meta sequence number.

Finally, because infinite mapping mode is significantly simpler, but
would require special casing all over the code, handle this by having a
separate sk_data_ready() callback for infinite mapping.  This also uses
tcp_read_sock(), but does away with the additional queue used for DSS
checksumming.

Signed-off-by: Doru Gucea <doru.gucea@intel.com>
Signed-off-by: Vlad Dogaru <vlad.dogaru@intel.com>