mptcp: Use correct tcp_queue when calling mptcp_fragment
The skb may come from the write-queue, the rtx-queue or the reinject
queue.

When it comes from the rtx-queue, we need to give mptcp_fragment the
right tcp_queue.

To handle the case when it comes from the reinject-queue, I will post a
follow-up patch as this seems to be a bug in the stable release as well.

Fixes: 8a1051d977cf ("Merge tag 'v4.15' into mptcp_trunk")
Signed-off-by: Christoph Paasch <cpaasch@apple.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
(cherry picked from commit edf4b8133e304238b1812c147455f65b2c15c436)
Signed-off-by: Christoph Paasch <cpaasch@apple.com>