mptcp: Remove cnt_established
We want to remove the number of up-calls into the meta-socket.
cnt_established is one of those where we are modifying a field in the
mpcb.

We can remove it fairly safely. It's only use is in some MPTCP-specific
congestion controls to short-circuit some parts of it.

One particular use was in mptcp_should_expand_sndbuf(). There, we use it
to see whether to expand the sndbuf when there are only backup subflows
available.

One problem with this old code is that the decision whether or not to send on
a backup subflow is steered by the modular scheduler. We cannot assume in
core-MPTCP code that backup subflows are never being sent on.

By no more taking backup subflows into account, we greatly simplify
mptcp_should_expand_sndbuf at the cost of potentially increasing the
window even if there is no subflow ready to send because it's a backup
subflow.

Signed-off-by: Christoph Paasch <cpaasch@apple.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
(cherry picked from commit 773f0e58474e56dfdf4334b889beebd4a8062d51)