skbuff: Add private data pointer
Allow struct sk_buff to be optionally extended when the control buffer
space is insufficient. New pointers for the extended data and associated
destructor are placed at the end of the data structure and left
uninitialized so cache behavior of struct sk_buff is unchanged when the
new pointers are not accessed.

Change-Id: I352466fbf7182b9e4570c7f820bea7f02a99f4b0
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>