mptcp: Remove IPv6 duplicated address detection
It causes us to take a refcnt on the address and actually periodically
schedule a timer until eventually the address gets either marked as
confirmed or removed.

It causes regressions in the Android test-suite.

This patch simply removes the duplicate address detection. That way,
when we get an IPv6-address we use it right away. Also, if the address
turns out to be duplicate on the link (which is really a cornercase
scenario), it will be removed and our subflow will be destroyed. I don't
think that's a big deal.

I don't suggest to backport this patch as it's more of a design-change
than a bugfix.

Github-issue: https://github.com/multipath-tcp/mptcp/issues/293

Signed-off-by: Christoph Paasch <cpaasch@apple.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
(cherry picked from commit 8e3743b348121ac66aba0df3074aba12c40591c2)
Signed-off-by: Christoph Paasch <cpaasch@apple.com>