ci:upd-tg: sync net-next with upstream
Upstream is here Davem's net-next.

Here in this script, we are going to only update the TopGit tree --
currently "empty" -- and not the "master" branch which originally
contains MPTCP implementation from multipath-tcp/mptcp.

For this first step, we only sync net-next branch with Davem's one. No
risk of error as we are not supposed to modify this branch, the base of
our tree.

Note that we also need to have a periodical sync to allow devs to
propose a change base on latest net-next: using the top of the tree --
t/upstream -- as target for inclusion in our tree or the base --
net-next -- for direct inclusion upstream like Florian did. We don't
want devs to update the tree manually.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Change-Id: I60de106b77de4d8f06524edf051c2f83f936d578