Update patch set 2
Patch Set 2:

(3 comments)

> Patch Set 2:
> 
> (3 comments)

Hi Mat,

Thank you for the review!

> I've seen git do some "clean" merges that turned out to have surprising consequences. Without both compilation and testing steps, I don't think automation should update the master branch.

I agree with you!

Note that here, I am going to only update the TopGit tree -- currently "empty" -- and not the "master" branch which originally contains MPTCP implementation from multipath-tcp/mptcp.
For this first step, I only sync net-next branch with Davem's one. No risk of error as we are not supposed to modify this branch, the base of our tree.

> We can still get a lot of value out of continuous integration by pushing to a secondary branch (like "ci-merge") and notifying when there are merge/build/test errors.

When the TG tree will be updated, I don't know what's the best:
- we locally update the entire tree and we reset in case of error
- we create a branch from the top of the tree and we merge with net-next (we might have more conflicts maybe?). Then in case of build/test error, it can be interesting to push this branch? But then it means it was also easy to generate this branch as the merge didn't create errors. So I don't know if it is useful to do that.

I don't know if it is interesting to push this ci-merge branch or a parallel TG tree then. I guess we don't want to do the merge manually after the push otherwise this job is not really valuable. If the merge + build + tests are successful, we can still track the changes easily like any public repo, no?

> Would this run periodically or triggered by net-next changes?

The goal would be to update the tree once per day, ideally only when net-next has been updated.

Note that we also need to have a periodical sync to allow devs to propose a change base on latest net-next: using the top of the tree -- t/upstream -- as target for inclusion in our tree or the base -- net-next -- for direct inclusion upstream like Florian did. We don't want devs to update the tree manually to be able to use 'git review <target>' here if you see what I mean :)

Patch-set: 2