Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> (1 comment)
> 
> Ok, I hadn't understood the proposed branch structure and ended off sending the discussion off track!

I will make sure to have a clearer description next time!

> If the goal is to just stay in sync with Dave's net-next, then I think fetching from davem/net-next.git and "git reset --hard FETCH_HEAD" would be a fine way to stay in sync. With "pull" being a "fetch + merge", if your workspace gets out of sync somehow (like if davem did a force push, which would be totally unexpected but not impossible) you want to be sure to still be mirroring the upstream repo rather than introducing merge commits. "git pull --ff-only" also works for this purpose.

Good idea, I just added '--ff-only'!

> Now that I better understand the purpose of the script, I agree that pushing the automated merge results doesn't seem valuable.

Once this is merged, I would propose to also update the rest of the tree and push the modifications if the compilation -- and later the tests -- are OK. Would it be OK for you to automate this as well? I guess once a patch is OK for our team, we will try to upstream it ASAP, right? If yes, the size of the tree should not be too big.

Patch-set: 4
Reviewer: Gerrit User 1015295 <1015295@d5d70762-12d0-45a1-890d-524b12d3f735>