mptcp: mptcp option len must be skb independent
Otherwise the TCP stack can't compute correctly the mss and
that will break segmentation.

Fix the issue reserving space for map if skb is NULL, too.
The map is effectively added according to skb extensions,
if available.
Overall we underestimate the MSS (always accounting for the map), but
we emit the map only if needed.

v2 -> v3:
 - reserve the map space if skb is NULL, as per Mat suggestion

v1 -> v2:
 - rebased on top of current git
 - use slightly different way to compute the opt len

Fixes: 9aaf99d52fc1 ("mptcp: Write MPTCP DSS headers to outgoing data packets")
Signed-off-by: Paolo Abeni <pabeni@redhat.com>
--
Notes:
- this can be squashed into the fixes commit
- the self-test still fails, there are more, unrelated, issues pending