ci:upd-tg: tag export branch with date
We might have references to commits from this 'export' branch. But this
branch is regularly rebuilt. Git GC would them remove these old
references.

Adding a tag force Github (and other clones) to keep these references.
All these tags will be like this:

    export/<date>

Note that if working from 'export' is an issue for you because it forces
you to use 'git rebase --onto', you can also work on top of 'for-review'
branch: the code will be the same, the git history will not be rewritten
for this branch.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Suggested-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Change-Id: I172661673bb515687c07471289a857f80e06b0c1