mptcp: selftests: increase timeout and return error on ping failure
Now that we add a qdisc that does random drop, there is a small chance
for the ping to be dropped, so this triggers:
ns2 -> 10.0.1.2 connectivity [ FAIL ]

To fix this, split the ping test part to its own loop and only
add the qdiscs after the ping tests have completed.

Also, the script would still exit with 0 status in this case, so update
run_tests to check the return value of do_transfer() rather than rely on
that function to set the global exit state in all cases.

Last, increase the timeout: with qdiscs present, 10 seconds is too low:
in some rare cases sender or receiver exited too early.

Signed-off-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>