mptcp: avoid validating MP_CAPABLE keys on 3way HS handling
The server side already stored the key on syn, and 3way HS ack
reception is fragile: the ack can be lost, and the following
data pkts processed instead.

This fixes the pending self-test issue in my testbed.

Suggested-by: Peter Krystad <peter.krystad@linux.intel.com>
Suggested-by: Christoph Paasch <cpaasch@apple.com>
Fixes: af550dd282ca ("mptcp: Create SUBFLOW socket for incoming connections")
Signed-off-by: Paolo Abeni <pabeni@redhat.com>
--
I'm ok with squashing into the Fixes commit, but that will likely cause
conflicts.

We can possible have [in the future] a similar issue with MP_JOIN, as we
validate the token on 3way HS ack reception, and that is fragile ??!

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>