mptcp: recvmsg() should really try to deliver 'len' bytes to user-space
Currently we break the recvmsg() main loop as soon as we have copied any
amount of data or the ssk is closed.

With reinjection we can receive duplicate data and reach the 'wait_for_data'
label with 'copied == 0'.
The subflow sk could be already in closed status, if the local socket
has already received all the data from the peer (and the data is sitting
into the receive queue).

When we meet the above scenario, currently recvmsg() terminates with
return code '0' and the user space interpret that as EoF, even if
there is potentially a lot of data pending.

This change addresses the issue checking for pending data before
going through the 'wait_for_data' code path.

v1 -> v2:
 - fix type in comment

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
--
I'm ok with squashing to commit 786f07cf97f1 ("mptcp: Implement MPTCP receive path")
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>