mptcp: allow MPTCP sockets by default
At LPC2019, the feedback was that it should be easy to create MPTCP
sockets to have testers. But still important to have ways to disable the
creation of new MPTCP sockets. It can be easily done via this new
sysctl, CGroups or SELinux. Netfilter can also be used to close existing
MPTCP connections if needed.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>