ci:upd-tg: option not to update the base
It is useful when we want to test changes only in the TopGit tree.

If the base is updated, we have to wait for the ci to finish to work on
the tree without having to re-sync after.
Also good if we want to check if a recent change on our side has broken
things (without having to suspect other changes from net-next).

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Change-Id: Ia54bc2b26c1314ee85112d40e58ecda7e52acf7c