selftests: mptcp: random ethtool tweaking
Instead of unconditionally disabling TSO in ns3, turn off any of gso/tso/gro
in ns3 and/or ns4.

This gets us various combinations of GRO/GSO/TSO without a large impact on
the average test run time.

The old approach is already not deterministic, and cannot possibly be
made so.  For protocol tests mptcp packetdrill is a better fit.

Signed-off-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>