selftests: mptcp: fix spurious deadlock
do_rnd_write() may call write on the mptcp socket repeatedly.
As the peerfd is blocking, this prevents the process from draining the
receive buffer.

Its unlikely but possible that both sides are blocked in write()
on the tcp socket, with both ends announcing a zero window,
waiting for read() to drain the socket receive buffer.

An alternative fix would be to make the peerfds nonblocking.
This change instead makes it so we call write at most once after
poll revents has the POLLOUT bit set.

Signed-off-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>