selftests: mptcp: remove the redundant test cases
Before this change we test all combinations, i.e.
ns1 -> ns{1,2,3,4}
ns2 -> ns{1,2,3,4}
ns3 -> ns{1,2,3,4}
ns4 -> ns{1,2,3,4}

and for each we test all of tcp -> mptcp, mptcp -> tcp and mptcp -> mptcp.

This means we have several redundant cases.
So only do ns1 -> ns1, but not ns2 -> ns2 and so on.

Also, the fallback test case is supposed to verify that we can still
connect, but there is no point in testing this any further.

This change makes it so that we *only* run mptcp -> mptcp tests,
and never repeat the loopback case (sender and receiver in same
netns).  Only exception: ns1->ns1, which also checks that mptcp sender
can connect to tcp receiver, and vice versa.

Signed-off-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>