selftests:mptcp: force MPTCP_IPV6 config
We want our stack to be tested with v6, better to force the kconfig
accordingly then.

Also if we are sure we have IPv6 support, we can simplify the script to
always setup the IPv6 config but not launch IPv6 tests. We can then now
have result in ~15/20 seconds with the "-4" option.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>